refactor CareEntry into new component
