+++ src/components/treeActivity/index.tsx
@@ -1,32 +1,13 @@
 import React, { useState } from 'react';
-import moment from 'moment';
-import {
-  Typography,
-  List,
-  Select,
-  Pagination,
-  Button,
-  Row,
-  Col,
-  Form,
-  Modal,
-} from 'antd';
-import { EditOutlined, CloseOutlined } from '@ant-design/icons';
+import { Typography, List, Select, Pagination } from 'antd';
 import {
   MonthYearOption,
   TreeCare,
 } from '../../containers/treePage/ducks/types';
-import { TitleProps } from 'antd/lib/typography/Title';
-import { DARK_GREEN, MID_GREEN, TEXT_GREY } from '../../utils/colors';
+import { DARK_GREEN, MID_GREEN } from '../../utils/colors';
 import { UNABBREVIATED_MONTHS } from '../../assets/content';
 import styled from 'styled-components';
-import StewardshipForm from '../forms/stewardshipForm';
-import { useSelector } from 'react-redux';
-import {
-  getPrivilegeLevel,
-  isAdmin,
-  getUserID,
-} from '../../auth/ducks/selectors';
+import CareEntry from '../careEntry';
 
 const TreeCareTitle = styled(Typography.Paragraph)`
   margin: 0px 5px;
@@ -34,26 +15,6 @@ const TreeCareTitle = styled(Typography.Paragraph)`
   font-weight: bold;
   line-height: 26px;
   color: ${DARK_GREEN};
-`;
-
-const CareEntry = styled.div`
-  margin: 15px;
-`;
-
-const EntryDate = styled(Typography.Paragraph)<TitleProps>`
-  display: inline;
-  text-align: center;
-  line-height: 0px;
-  font-size: 18px;
-  font-weight: bold;
-  color: ${DARK_GREEN};
-`;
-
-const EntryMessage = styled(Typography.Paragraph)`
-  display: inline;
-  text-align: center;
-  line-height: 0px;
-  color: ${TEXT_GREY};
 `;
 
 const StewardshipActivityDropdownContainer = styled.div`
@@ -73,24 +34,6 @@ const CenteredPagination = styled(Pagination)`
   }
   & .ant-pagination-item-active :hover {
     background-color: ${DARK_GREEN};
-  }
-`;
-
-const EditButton = styled(Button)`
-  color: white;
-  float: right;
-  font-size: 20px;
-  padding: 0px 10px;
-  line-height: 0px;
-`;
-
-const StyledClose = styled(CloseOutlined)`
-  color: red;
-  padding: 5px;
-  border-radius: 3px;
-
-  & :hover {
-    background-color: #ffd1d1;
   }
 `;
 
@@ -128,25 +71,6 @@ const TreeActivity: React.FC<TreeActivityProps> = ({
 
   const [pageNumber, setPageNumber] = useState(0);
 
-  const [editActivityId, setEditActivityId] = useState<number>(undefined);
-  const [stewardshipFormInstance] = Form.useForm();
-
-  const userIsAdmin: boolean = useSelector((state: C4CState) =>
-    isAdmin(state.authenticationState.tokens),
-  );
-
-  const userId: number = useSelector((state: C4CState) =>
-    getUserID(state.authenticationState.tokens),
-  );
-
-  function showButtons(activityOwner) {
-    return userIsAdmin || userId == activityOwner;
-  }
-
-  function treeCareToMoment(month: string, day: string, year: number) {
-    return moment(`${month} ${day} ${year}`, 'MMM Do YYYY');
-  }
-
   return (
     <>
       {/* {console.log(selectOptions)} */}
@@ -176,48 +100,11 @@ const TreeActivity: React.FC<TreeActivityProps> = ({
           emptyText: 'No Stewardship Activities Recorded for this Tree',
         }}
         renderItem={(activity, key) => (
-          <CareEntry key={key}>
-            <Row>
-              <Col span={5}>
-                <EntryDate>{activity.month + ' ' + activity.day}</EntryDate>
-              </Col>
-              <Col span={1} />
-              <Col span={18}>
-                <EntryMessage>{activity.message}</EntryMessage>
-                {showButtons(activity.ownerId) && (
-                  <>
-                    <EditButton
-                      type="primary"
-                      onClick={() => setEditActivityId(activity.activityId)}
-                    >
-                      <EditOutlined />
-                    </EditButton>
-                    <Modal
-                      bodyStyle={{ paddingBottom: '5px' }}
-                      title="Edit stewardship"
-                      visible={editActivityId === activity.activityId}
-                      onCancel={() => setEditActivityId(undefined)}
-                      footer={null}
-                      closeIcon={<StyledClose />}
-                    >
-                      <StewardshipForm
-                        onFinish={onFinishEditStewardship(
-                          activity.activityId,
-                          stewardshipFormInstance,
-                        )}
-                        form={stewardshipFormInstance}
-                        initialDate={treeCareToMoment(
-                          activity.month,
-                          activity.day,
-                          activity.year,
-                        )}
-                      />
-                    </Modal>
-                  </>
-                )}
-              </Col>
-            </Row>
-          </CareEntry>
+          <CareEntry
+            key={key}
+            activity={activity}
+            onFinishEditStewardship={onFinishEditStewardship}
+          />
         )}
       />
       <CenteredPagination
